Moved AssimpResourceAdaptor to dynamics.
